requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
ffmpeg==1.4
filelock==3.0.12
funcsigs==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
imageio==2.5.0
importlib-metadata==0.23
ipydatawidgets==4.0.1
ipykernel==5.1.2
ipympl==0.3.3
ipython==7.8.0
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipywebrtc==0.5.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.16.1
opencv-python==4.1.1.26
packaging==19.2
pandas==0.24.1
pandocfilters==1.4.2
parso==0.5.1
pathlib2==2.3.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pipenv==2018.11.26
pluggy==0.13.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.2
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pymap3d~=2.4.3
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.3
python-dateutil==2.8.0
pythreejs==2.1.1
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
ray==0.7.5
redis==3.3.8
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
setproctitle==1.1.10
six==1.12.0
sklearn==0.0
smmap2==2.0.5
tensorboardX==1.8
terminado==0.8.2
testpath==0.4.2
torch==1.2.0
torchvision==0.4.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
traittypes==0.2.1
urllib3==1.25.6
virtualenv==16.7.5
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
git+git://github.com/matplotlib/basemap.git